+++ src/main/java/pl/edu/agh/miss/intruders/api/Config.java
@@ -4,5 +4,5 @@ package pl.edu.agh.miss.intruders.api;
  * Created by tgs on 3/20/16.
  */
 public interface Config {
-    int getNewProbability(int currentProbability, Robot robot);
+    float getNewProbability(float currentProbability, Robot robot);
 }
